# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: MetalMark API
  description: Self-hosted personal finance (Monarch-style). LAN/VPN only.
  version: 0.1.0
paths:
...
  /auth/signup:
    post:
      tags:
      - auth
      summary: Signup
      operationId: signup_auth_signup_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /auth/login:
    post:
      tags:
      - auth
      summary: Login
      operationId: login_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /auth/logout:
    post:
      tags:
      - auth
      summary: Logout
      operationId: logout_auth_logout_post
      parameters:
      - name: metalmark_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Metalmark Session
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /auth/me:
    get:
      tags:
      - auth
      summary: Me
      operationId: me_auth_me_get
      parameters:
      - name: metalmark_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Metalmark Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /accounts:
    post:
      tags:
      - accounts
      summary: Create Account
      operationId: create_account_accounts_post
      parameters:
      - name: metalmark_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Metalmark Session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - accounts
      summary: List Accounts
      operationId: list_accounts_accounts_get
      parameters:
      - name: owner_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Owner Id
      - name: metalmark_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Metalmark Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AccountOut'
                title: Response List Accounts Accounts Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /accounts/net-worth:
    get:
      tags:
      - accounts
      summary: Net Worth
      operationId: net_worth_accounts_net_worth_get
      parameters:
      - name: owner_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Owner Id
      - name: metalmark_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Metalmark Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NetWorthOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /accounts/{account_id}:
    get:
      tags:
      - accounts
      summary: Get Account
      operationId: get_account_accounts__account_id__get
      parameters:
      - name: account_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Account Id
      - name: metalmark_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Metalmark Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - accounts
      summary: Update Account
      operationId: update_account_accounts__account_id__patch
      parameters:
      - name: account_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Account Id
      - name: metalmark_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Metalmark Session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - accounts
      summary: Delete Account
      operationId: delete_account_accounts__account_id__delete
      parameters:
      - name: account_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Account Id
      - name: metalmark_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Metalmark Session
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /household:
    get:
      tags:
      - household
      summary: Get Household
      operationId: get_household_household_get
      parameters:
      - name: metalmark_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Metalmark Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HouseholdOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - household
      summary: Update Household
      description: 'Owner-only: the household name is what every member sees as their account.'
      operationId: update_household_household_patch
      parameters:
      - name: metalmark_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Metalmark Session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HouseholdUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HouseholdOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /household/members:
    get:
      tags:
      - household
      summary: List Members
      operationId: list_members_household_members_get
      parameters:
      - name: metalmark_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Metalmark Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MemberOut'
                title: Response List Members Household Members Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    AccountCreate:
      properties:
        name:
          type: string
          maxLength: 200
          minLength: 1
          title: Name
        type:
          type: string
          pattern: ^(depository|credit|investment|loan|other)$
          title: Type
        currency:
          type: string
          maxLength: 3
          minLength: 3
          title: Currency
        subtype:
          anyOf:
          - type: string
          - type: 'null'
          title: Subtype
        institution:
          anyOf:
          - type: string
          - type: 'null'
          title: Institution
        current_balance:
          anyOf:
          - type: number
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Current Balance
          default: '0'
        balance_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Balance Date
        owner_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Owner Id
      type: object
      required:
      - name
      - type
      - currency
      title: AccountCreate
    AccountOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        name:
          type: string
          title: Name
        type:
          type: string
          title: Type
        currency:
          type: string
          title: Currency
        subtype:
          anyOf:
          - type: string
          - type: 'null'
          title: Subtype
        institution:
          anyOf:
          - type: string
          - type: 'null'
          title: Institution
        current_balance:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Current Balance
        balance_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Balance Date
        is_asset:
          type: boolean
          title: Is Asset
        owner_id:
          type: string
          format: uuid
          title: Owner Id
        is_manual:
          type: boolean
          title: Is Manual
        is_hidden:
          type: boolean
          title: Is Hidden
      type: object
      required:
      - id
      - name
      - type
      - currency
      - subtype
      - institution
      - current_balance
      - balance_date
      - is_asset
      - owner_id
      - is_manual
      - is_hidden
      title: AccountOut
    AccountUpdate:
      properties:
        name:
          anyOf:
          - type: string
            maxLength: 200
            minLength: 1
          - type: 'null'
          title: Name
        subtype:
          anyOf:
          - type: string
          - type: 'null'
          title: Subtype
        institution:
          anyOf:
          - type: string
          - type: 'null'
          title: Institution
        current_balance:
          anyOf:
          - type: number
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          - type: 'null'
          title: Current Balance
        balance_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Balance Date
        owner_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Owner Id
        is_hidden:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Hidden
      type: object
      title: AccountUpdate
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    HouseholdOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        name:
          type: string
          title: Name
        base_currency:
          type: string
          title: Base Currency
        timezone:
          type: string
          title: Timezone
        role:
          type: string
          title: Role
      type: object
      required:
      - id
      - name
      - base_currency
      - timezone
      - role
      title: HouseholdOut
    HouseholdUpdate:
      properties:
        name:
          anyOf:
          - type: string
            maxLength: 200
            minLength: 1
          - type: 'null'
          title: Name
        timezone:
          anyOf:
          - type: string
            maxLength: 64
            minLength: 1
          - type: 'null'
          title: Timezone
      additionalProperties: false
      type: object
      title: HouseholdUpdate
      description: '``base_currency`` is absent on purpose: it is immutable once set (ADR-0017).


        ``extra="forbid"`` makes that absence enforceable: without it a PATCH carrying

        ``base_currency`` would be accepted and silently ignored, and the caller would

        believe the currency had changed.'
    LoginRequest:
      properties:
        email:
          type: string
          format: email
          title: Email
        password:
          type: string
          title: Password
      type: object
      required:
      - email
      - password
      title: LoginRequest
    MeResponse:
      properties:
        user:
          $ref: '#/components/schemas/UserResponse'
        household_id:
          type: string
          format: uuid
          title: Household Id
        household_name:
          type: string
          title: Household Name
        base_currency:
          type: string
          title: Base Currency
        role:
          type: string
          title: Role
        csrf_token:
          type: string
          title: Csrf Token
      type: object
      required:
      - user
      - household_id
      - household_name
      - base_currency
      - role
      - csrf_token
      title: MeResponse
    MemberOut:
      properties:
        user_id:
          type: string
          format: uuid
          title: User Id
        display_name:
          type: string
          title: Display Name
        email:
          type: string
          format: email
          title: Email
        role:
          type: string
          title: Role
      type: object
      required:
      - user_id
      - display_name
      - email
      - role
      title: MemberOut
    NetWorthOut:
      properties:
        base_currency:
          type: string
          title: Base Currency
        assets:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Assets
        liabilities:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Liabilities
        net_worth:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Net Worth
        unconverted_currencies:
          items:
            type: string
          type: array
          title: Unconverted Currencies
          default: []
        attribution:
          type: string
          title: Attribution
          default: account
      type: object
      required:
      - base_currency
      - assets
      - liabilities
      - net_worth
      title: NetWorthOut
...
    SignupRequest:
      properties:
        email:
          type: string
          format: email
          title: Email
        display_name:
          type: string
          maxLength: 120
          minLength: 1
          title: Display Name
        password:
          type: string
          maxLength: 256
          minLength: 8
          title: Password
        household_name:
          anyOf:
          - type: string
            maxLength: 200
            minLength: 1
          - type: 'null'
          title: Household Name
      type: object
      required:
      - email
      - display_name
      - password
      title: SignupRequest
...
    UserResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        email:
          type: string
          format: email
          title: Email
        display_name:
          type: string
          title: Display Name
        is_admin:
          type: boolean
          title: Is Admin
      type: object
      required:
      - id
      - email
      - display_name
      - is_admin
      title: UserResponse
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError